backend/requirements.txt
aniso8601==9.0.1
black==22.12.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
google-auth==2.16.0
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.46
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.2.2
